% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_of_death}
\alias{age_of_death}
\title{Probability death with be at a given age.}
\format{A vector with 90 probabilities.}
\source{
See create lifetable data for more information
}
\usage{
age_of_death
}
\description{
Estimated from Mali lifetable data.
}
\keyword{datasets}
