% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{amprod}
\alias{amprod}
\title{Array-matrix product (R implementation)}
\usage{
amprod(A, M, k)
}
\description{
This R implementation is used for operations not covered by the specialized Cpp optimizations (like 4D Z).
}
\keyword{internal}
