% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{attack_rate}
\alias{attack_rate}
\title{Attack rate}
\usage{
attack_rate(sim)
}
\arguments{
\item{sim}{An object of class \code{"sim_epi"}.}
}
\value{
A numeric scalar giving the attack rate.
}
\description{
Computes the cumulative number of infections over the course of the epidemic,
defined as the integral (sum) of the incidence curve.
}
\details{
Let \eqn{\lambda(t)} denote the incidence function.
The attack rate is defined as
\deqn{
AR = \int_0^T \lambda(t)\, dt.
}

In discrete time simulations, this integral is approximated by
\deqn{
AR \approx \sum_{t=0}^T \lambda(t).
}

For models with permanent immunity, this corresponds to the final epidemic
size.

This metric requires a model flow named \code{"incidence"}.
}
\examples{
sim <- simulate_epi(
  model = SIR_MODEL,
  times = 0:160,
  parms = c(beta = 0.3, gamma = 0.1),
  init  = c(S = 999, I = 1, R = 0)
)

attack_rate(sim)

}
