% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{data}
\name{cbPalette}
\alias{cbPalette}
\title{Color-blind friendly color palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
cbPalette
}
\description{
Color-blind friendly color palette
}
\keyword{datasets}
