% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\docType{data}
\name{cc.genes_mmusculus}
\alias{cc.genes_mmusculus}
\title{cc.genes_mmusculus}
\format{
An object of class \code{list} of length 2.
}
\usage{
cc.genes_mmusculus
}
\description{
cc.genes_mmusculus
}
\keyword{datasets}
