% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_equilibrium}
\alias{check_equilibrium}
\title{Check Distance from Equilibrium}
\usage{
check_equilibrium(parameters, data)
}
\arguments{
\item{parameters}{system parameters}

\item{data}{value of each population}
}
\value{
data.frame of maximum error of each derivative in the reinfection
model, each derivative evaluated at the data, and the derivative of
the total population
}
\description{
This function computes how close the provided data is to an equilibrium
}
