% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_populations}
\alias{combine_populations}
\title{Combine Populations}
\usage{
combine_populations(solution = NULL, type = "reinfection")
}
\arguments{
\item{solution_list}{list of data.frames to combine into a single data.frame}
}
\value{
combined data.frame
}
\description{
This function combines the data.frames provided in the list, while adding
a "path" column to distinguish the data.frame data.  This can be used to
combine simulation runs to generate a 3D plot with multiple paths.
}
