% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compute_equilibrium_error}
\alias{compute_equilibrium_error}
\title{Compute Equilibrium Error}
\usage{
compute_equilibrium_error(parameters, data)
}
\arguments{
\item{parameters}{system parameters}

\item{data}{value of each population}
}
\value{
maximum of absolute value of each derivative evaluated at data
}
\description{
This function computes the max error of an point to being an equilibrium
}
