% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilbrium_birth_death_sampling.R
\name{conditional_death_age}
\alias{conditional_death_age}
\title{Draw conditional lifespan}
\usage{
conditional_death_age(age, age_of_death, max_age = 89)
}
\arguments{
\item{age}{Current age (days)}

\item{max_age}{Maximum age (in years, actual maximum is this + 364 days)}

\item{equilbrium_age}{Equilirium probability of death}
}
\value{
A single conditioned lifespan draw
}
\description{
Makes a draw of the lifespan (in days), given a current age
}
