% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_birth_death_sampling.R
\name{conditional_lifespan}
\alias{conditional_lifespan}
\title{Draw conditional lifespan}
\usage{
conditional_lifespan(age, age_of_death, max_age = 89)
}
\arguments{
\item{age}{Current age (days)}

\item{age_of_death}{Equilibrium probability of death}

\item{max_age}{Maximum age (in years, actual maximum is this + 364 days)}
}
\value{
A single conditioned lifespan draw
}
\description{
Makes a draw of the lifespan (in days), given a current age
}
