% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_construct_Walpha_design_dyn}
\alias{cpp_construct_Walpha_design_dyn}
\title{Construct Design Matrix for Beta Updates with Dynamic W}
\usage{
cpp_construct_Walpha_design_dyn(W_field, X, alpha)
}
\arguments{
\item{W_field}{List of T cubes, each m x m x p (one per time period).}

\item{X}{Three-dimensional array (m x m x T) of network states.}

\item{alpha}{Vector (p x 1) of current sender parameters.}
}
\value{
Matrix (m*m*T x p) design matrix for beta GLM step.
}
\description{
Builds the design matrix for updating receiver effects when influence
covariates vary over time (4D W array: m x m x p x T).
}
