% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_construct_Wbeta_design_dyn}
\alias{cpp_construct_Wbeta_design_dyn}
\title{Construct Design Matrix for Alpha Updates with Dynamic W}
\usage{
cpp_construct_Wbeta_design_dyn(W_field, X, beta)
}
\arguments{
\item{W_field}{List of T cubes, each m x m x p (one per time period).}

\item{X}{Three-dimensional array (m x m x T) of network states.}

\item{beta}{Vector (p x 1) of current receiver parameters.}
}
\value{
Matrix (m*m*T x p) design matrix for alpha GLM step.
}
\description{
Builds the design matrix for updating sender effects when influence
covariates vary over time (4D W array: m x m x p x T).
}
