% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mll_gH_dyn}
\alias{cpp_mll_gH_dyn}
\title{Calculate Gradient and Hessian with Dynamic W}
\usage{
cpp_mll_gH_dyn(tab, Y, W_field, X, Z_list, family)
}
\arguments{
\item{tab}{Parameter vector [theta, alpha_2:p, beta].}

\item{Y}{Three-dimensional array (m x m x T) of outcomes.}

\item{W_field}{List of T cubes, each m x m x p.}

\item{X}{Three-dimensional array (m x m x T).}

\item{Z_list}{List of q cubes (m x m x T), one per covariate.}

\item{family}{Distribution family string.}
}
\value{
List with grad, hess, shess (after identifiability projection).
}
\description{
Computes gradient and Hessian of the negative log-likelihood when
influence covariates W vary over time (4D: m x m x p x T). The W
array is passed as a list of cubes to avoid Rcpp 4D array limitations.
}
