% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_birth_death_sampling.R
\name{current_age}
\alias{current_age}
\title{Draw current age}
\usage{
current_age(N, equilbrium_age)
}
\arguments{
\item{N}{Number of people}

\item{equilbrium_age}{Equilirium age distribution}
}
\value{
A vector of N age draws
}
\description{
Make draw(s) of current age (in days) at equilbrium
}
