% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{days_to_steps}
\alias{days_to_steps}
\title{Convert days to steps}
\usage{
days_to_steps(days, substep)
}
\arguments{
\item{days}{Number of days}

\item{substep}{Substeps per day}
}
\description{
Convert days to steps
}
