% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{demog_test}
\alias{demog_test}
\title{Test demography}
\usage{
demog_test(N, days, substep, age_of_death, equilibrium_age)
}
\arguments{
\item{N}{Population size}

\item{days}{Simulation time (days)}

\item{substep}{Substeps per day}

\item{age_of_death}{Probability of age of death}

\item{equilibrium_age}{Equilibrium age distribution}
}
\description{
Test demography
}
