% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_equilibrium_age}
\alias{draw_equilibrium_age}
\title{Draw an age (in days) from the equilibrium age distribution}
\usage{
draw_equilibrium_age(age_years, equilibrium_age)
}
\arguments{
\item{equilibrium_age}{Equilribrium age distribution}

\item{age_year}{Vector of integersof possible age years}
}
\value{
An integer age (days)
}
\description{
Draw an age (in days) from the equilibrium age distribution
}
