% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_equilibrium_death_age}
\alias{draw_equilibrium_death_age}
\title{Draw a death age conditional on an age from the equilibrium age distribution}
\usage{
draw_equilibrium_death_age(age, age_years, age_of_death, max_age)
}
\arguments{
\item{age}{Age (days) drawn from equilibrium age distribution}

\item{age_years}{Vector of integers of possible age years}

\item{age_of_death}{Equilribrium age of death distribution}

\item{max_age}{Maximum age (years)}
}
\value{
An integer death age (days)
}
\description{
Draw a death age conditional on an age from the equilibrium age distribution
}
