% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM_closed_SIR.R
\name{draw_sample}
\alias{draw_sample}
\title{Sample events for a group of individuals from binomial}
\usage{
draw_sample(people, prob)
}
\arguments{
\item{people}{index of individuals}

\item{prob}{Binomial probability}
}
\value{
Sampled successes from people
}
\description{
Sample events for a group of individuals from binomial
}

