% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equilibrium_age}
\alias{equilibrium_age}
\title{Equilibrium age distribution.}
\format{A vector with 90 probabilities.}
\source{
See create lifetable data for more information
}
\usage{
equilibrium_age
}
\description{
Estimated from Mali lifetable data.
}
\keyword{datasets}
