% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_demography.R
\name{equilibrium_age_death}
\alias{equilibrium_age_death}
\title{Equilibrium age of death}
\usage{
equilibrium_age_death(life_table)
}
\arguments{
\item{life_table}{Vector of yearly death probability for 0-90.}
}
\description{
Equilibrium age of death
}
