% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_demography.R
\name{equilibrium_age_distribution}
\alias{equilibrium_age_distribution}
\title{Equilibrium age distribution}
\usage{
equilibrium_age_distribution(life_table)
}
\arguments{
\item{life_table}{Vector of yearly death probability for 0-90.}
}
\value{
Proportion of popution in each age group (at equilibrium)
}
\description{
Estimate the equilbrium age distribution given a lifetable (probability, of death in each year).
Life table raw data can be obtained from http://apps.who.int/gho/data/node.main.687?lang=en. The input
used is the "nqx - probability of dying between ages x and x+n" column, which is then smoothly
extrapolated to yearly probabilities.
}
