% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_Z}
\alias{flatten_Z}
\title{Flatten Z for GLM input}
\usage{
flatten_Z(Z)
}
\arguments{
\item{Z}{(m x m x q x T) array.}
}
\value{
(m*m*T) x q matrix.
}
\description{
Converts Z (m x m x q x T) into a (m*m*T) x q matrix, matching the flattening of Y.
}
