% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_sir_wrapper.R
\name{format_cpp}
\alias{format_cpp}
\title{Format c++}
\usage{
format_cpp(out)
}
\arguments{
\item{out}{Raw model output}
}
\value{
Formated output
}
\description{
Formats c++ model output (column order and selection)
}
