% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{gene_name_to_ensembl}
\alias{gene_name_to_ensembl}
\title{convert gene names to ensemble ids using biomarT}
\usage{
gene_name_to_ensembl(
  external_gene_names,
  mart = useEnsembl(biomart = "ensembl", dataset = "mmusculus_gene_ensembl", GRCh = 37)
)
}
\arguments{
\item{external_gene_names}{character vector of gene names to convert}
}
\description{
convert gene names to ensemble ids using biomarT
}
