% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_base_plot}
\alias{generate_base_plot}
\title{Generate Base Plot}
\usage{
generate_base_plot(data = NULL)
}
\arguments{
\item{data}{data source for visualization (if not null)}
}
\value{
An empty plotly line plot object
}
\description{
This function generates an empty plotly line plot.
}
