% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_initial_conditions}
\alias{generate_initial_conditions}
\title{Generate Initial Conditions}
\usage{
generate_initial_conditions()
}
\value{
initial condition as a list
}
\description{
This function generates an initial condition of the form
(S1, S2, I1, I2, A1, A2, R1=0, R2=0), where S1, S2, I1, I2, A1, and A2 are
each selected from a uniform distribution on \link{0,100} with the constraint
that the sum is 100.
}
