% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_population_plot}
\alias{generate_population_plot}
\title{Generate Population Plot}
\usage{
generate_population_plot(data = NULL)
}
\arguments{
\item{data}{data source for visualization (if not null)}
}
\value{
An empty plotly line plot object
}
\description{
This function is a wrapper around the generate_base_plot function which
defaults the y_limits parameter to display between 0 and 100.
}
