% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_tenery_plot}
\alias{generate_tenery_plot}
\title{Generate Tenery Plot}
\usage{
generate_tenery_plot()
}
\value{
An tenery plot object
}
\description{
This function creates a tenery plotly plot object.
}
