% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_color_palette}
\alias{get_color_palette}
\title{Get Color Palette}
\usage{
get_color_palette()
}
\value{
list of colors in color-blind friendly color palette
}
\description{
This function returns the color-blind friendly color palette.
}
