% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_epi_model}
\alias{get_epi_model}
\title{Get a registered epidemiological model}
\usage{
get_epi_model(name)
}
\arguments{
\item{name}{Model name}
}
\description{
Get a registered epidemiological model
}
