% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_sir_wrapper.R
\name{get_time}
\alias{get_time}
\title{Model times}
\usage{
get_time(days, substep)
}
\arguments{
\item{days}{Number of days}

\item{substep}{Steps per day}
}
\value{
Vector of times
}
\description{
Model times
}
