% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM_closed_SIR.R
\name{ibm_closed_sir}
\alias{ibm_closed_sir}
\title{Closed SIR (IBM)}
\usage{
ibm_closed_sir(sigma = 2, beta = 6, N = 1000, i0 = 1, days = 100,
  substep = 1)
}
\arguments{
\item{sigma}{Recovery rate (1 / mean duration of infection)}

\item{beta}{Transmission parameter}

\item{N}{Population size}

\item{i0}{Number of infected people at time = 0}

\item{days}{End time}

\item{substep}{time step}
}
\value{
IBM closed SIR model output
}
\description{
Closed SIR (IBM)
}
