% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ibm_closed_sir_rcpp}
\alias{ibm_closed_sir_rcpp}
\title{ibm_closed_sir_rcpp 1}
\usage{
ibm_closed_sir_rcpp(sigma = 2, beta = 4, N = 1000L, i0 = 1L,
  t_final = 100L, dt = 0.01)
}
\arguments{
\item{sigma}{Recovery rate (1 / mean duration of infection)}

\item{beta}{Transmission parameter}

\item{N}{Population size}

\item{i0}{Number of infected people at time = 0}

\item{t_final}{End time}

\item{dt}{time step}
}
\value{
IBM model output
}
\description{
ibm_closed_sir_rcpp 1
}
