% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infect_corners.R
\name{infect_corners}
\alias{infect_corners}
\title{Infecting the corners of a given matrix}
\usage{
infect_corners(nrow = 20, ncol = 20)
}
\arguments{
\item{nrow}{An integer indicating number of rows of expected matrix}

\item{ncol}{An integer indicating number of columns of expected matrix}
}
\value{
A matrix of special class: SIRmatrix
}
\description{
Infecting the corners of a given matrix
}
\examples{
infect_corners()

infect_corners(nrow = 50, ncol = 50)
}
