% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM_closed_SIR.R
\name{infection_probability}
\alias{infection_probability}
\title{Calculate the current probability of infection}
\usage{
infection_probability(beta, I_prev, N_inv, dt)
}
\arguments{
\item{beta}{Transmission parameter}

\item{I_prev}{Number of infected people in previous timestep}

\item{N_inv}{Inverse of the number of people}

\item{dt}{time step}
}
\value{
Probability of infection
}
\description{
Calculate the current probability of infection
}

