% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_step.R
\name{infection_step}
\alias{infection_step}
\title{A single infection time step}
\usage{
infection_step(x, prob)
}
\arguments{
\item{x}{A matrix of 0's, 1's, and/or 2's.}

\item{prob}{A float value of the probability of infection of a given cell.}
}
\value{
A matrix of special class: SIRmatrix
}
\description{
A single infection time step
}
\examples{
infection_step(infect_corners(30, 30), 0.2)

infection_step(random_infection(30, 30), 0.1)
}
