% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{initial_doubling_time}
\alias{initial_doubling_time}
\title{Initial doubling time}
\usage{
initial_doubling_time(sim)
}
\arguments{
\item{sim}{An object of class \code{"sim_epi"}.}
}
\value{
A numeric scalar giving the initial doubling time.
}
\description{
Computes the epidemic doubling time during the initial exponential growth
phase.
}
\details{
If incidence grows exponentially, the doubling time is
\deqn{
T_d = \frac{\log 2}{r}.
}
}
\examples{
sim <- simulate_epi(
  model = SI_MODEL,
  times = 0:50,
  parms = c(beta = 0.4),
  init  = c(S = 999, I = 1)
)

initial_doubling_time(sim)

}
