% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{initial_growth_rate}
\alias{initial_growth_rate}
\title{Initial exponential growth rate}
\usage{
initial_growth_rate(incidence, time, n = 7)
}
\arguments{
\item{incidence}{Numeric vector of incidence values.}

\item{time}{Numeric vector of time points.}

\item{n}{Integer. Number of initial time points used for estimation.}
}
\value{
A numeric scalar giving the estimated growth rate.
}
\description{
Estimates the early exponential growth rate by fitting a log-linear model
to the initial segment of an incidence curve.
}
\details{
Assuming exponential growth:
\deqn{
\lambda(t) \approx C e^{r t}.
}

Taking logarithms:
\deqn{
\log \lambda(t) = \log C + r t.
}
}
