% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{initial_infect}
\alias{initial_infect}
\title{Create a randomly infected matrix}
\usage{
initial_infect(num_rows = 3, num_cols = 5, p_infect = 0.1)
}
\arguments{
\item{num_rows}{number of rows in grid}

\item{num_cols}{number of columns in grid}

\item{p_infect}{probability of infection}
}
\description{
Create a randomly infected matrix
}
