% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_Submission.R
\name{is_infectious}
\alias{is_infectious}
\title{Returns true if current matrix contains infections; otherwise false.}
\usage{
is_infectious(x)
}
\arguments{
\item{x}{a matrix with susceptible, infected, and removed values}
}
\description{
Returns true if current matrix contains infections; otherwise false.
}
