% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_demography.R
\name{lifetable_to_transtion_matrix}
\alias{lifetable_to_transtion_matrix}
\title{Lifetable to transition matrix}
\usage{
lifetable_to_transtion_matrix(life_table)
}
\arguments{
\item{life_table}{Vector of yearly death probability for 0-90.}
}
\value{
Transition matrix
}
\description{
Lifetable to transition matrix
}
