% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mll_sir}
\alias{mll_sir}
\title{Calculate Negative Log-Likelihood}
\usage{
mll_sir(tab, Y, W, X, Z, family)
}
\arguments{
\item{tab}{Parameter vector.}

\item{Y}{Outcome array.}

\item{W}{Influence covariates.}

\item{X}{Bilinear covariates.}

\item{Z}{Exogenous covariates.}

\item{family}{Distribution family.}
}
\value{
The negative log-likelihood value.
}
\description{
Calculate Negative Log-Likelihood
}
