% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{open_sir_rcpp}
\alias{open_sir_rcpp}
\title{Test demography}
\usage{
open_sir_rcpp(N, days, substep, age_of_death, equilibrium_age, beta, sigma,
  i0)
}
\arguments{
\item{N}{Population size}

\item{days}{Simulation time (days)}

\item{substep}{Substeps per day}

\item{age_of_death}{Probability of age of death}

\item{equilibrium_age}{Equilibrium age distribution}

\item{beta}{Beta parameter (per day)}

\item{sigma}{Recovery rate (per day)}

\item{i0}{Initial number of infecteds}
}
\description{
Test demography
}
