% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{peak_incidence}
\alias{peak_incidence}
\title{Peak incidence}
\usage{
peak_incidence(incidence, time)
}
\arguments{
\item{incidence}{Numeric vector giving the incidence curve.}

\item{time}{Numeric vector of the same length giving time points.}
}
\value{
A named list with elements \code{time} and \code{value}.
}
\description{
Computes the maximum value of an incidence curve and the time at which
it occurs.
}
\details{
Let \eqn{\lambda(t)} denote the incidence function.
The peak incidence is defined as
\deqn{
\lambda_{\max} = \max_t \lambda(t),
}
and the time to peak incidence is
\deqn{
t^* = \arg\max_t \lambda(t).
}
}
\examples{
inc  <- c(1, 3, 7, 5, 2)
time <- 0:4
peak_incidence(inc, time)

}
