% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{peak_prevalence}
\alias{peak_prevalence}
\title{Peak prevalence}
\usage{
peak_prevalence(prevalence)
}
\arguments{
\item{prevalence}{Numeric vector representing prevalence over time.}
}
\value{
A numeric scalar giving the peak prevalence.
}
\description{
Computes the maximum value of a prevalence curve.
}
\details{
Let \eqn{I(t)} denote the prevalence over time.
The peak prevalence is defined as
\deqn{
I_{\max} = \max_t I(t).
}
}
\examples{
I <- c(1, 5, 8, 4, 2)
peak_prevalence(I)

}
