% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{plot.epi_model_predict}
\alias{plot.epi_model_predict}
\title{Plot predictions from a fitted epidemic model}
\usage{
\method{plot}{epi_model_predict}(x, type_obs = "h", col_obs = "black", col_pred = "red", lwd_pred = 2, ...)
}
\arguments{
\item{x}{An object of class \code{"epi_model_predict"}.}

\item{type_obs}{Plot type for observed data (default: \code{"h"}).}

\item{col_obs}{Color for observed incidence.}

\item{col_pred}{Color for predicted incidence.}

\item{lwd_pred}{Line width for predicted incidence.}

\item{...}{Further graphical parameters passed to \code{\link{plot}}.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Plot method for objects of class \code{"epi_model_predict"}.
The function displays the observed incidence data used for fitting together
with the model-predicted incidence trajectory.
}
