% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SIRmatrix.R
\name{plot.SIRmatrix}
\alias{plot.SIRmatrix}
\title{Wrapper plotting function for class SIRmatrix}
\usage{
\method{plot}{SIRmatrix}(x, title = "Infection", ...)
}
\arguments{
\item{x}{A matrix of 0's, 1's, and/or 2's.}

\item{title}{A string for the title of the plot.}

\item{...}{Any additional parameters that will be passed onto image().}
}
\value{
A generated plot using image()
}
\description{
Wrapper plotting function for class SIRmatrix
}
\examples{
plot(infect_corners())

plot(random_infection(), title = "Initial Random Infection")
}
