% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_Submission.R
\name{plot_infection}
\alias{plot_infection}
\title{Plot a matrix}
\usage{
plot_infection(dd)
}
\arguments{
\item{dd}{a matrix with susceptible, infected, and removed values}
}
\description{
Plot a matrix
}
