% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_open_sir_wrapper.R
\name{post}
\alias{post}
\title{Post processing wrapper for c++ output}
\usage{
post(out, t, step)
}
\arguments{
\item{out}{Raw model output}
}
\value{
Processed c++ output
}
\description{
Post processing wrapper for c++ output
}
