% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.fit_epi_model}
\alias{predict.fit_epi_model}
\title{Predict epidemic dynamics from a fitted epidemic model}
\usage{
\method{predict}{fit_epi_model}(object, n_days, init = NULL, times = NULL, method = "lsoda", ...)
}
\arguments{
\item{object}{An object of class \code{"fit_epi_model"}.}

\item{n_days}{Integer. Number of days to predict ahead.}

\item{init}{Named numeric vector of initial states. If \code{NULL}, the initial
state is reconstructed using \code{object$ini0}.}

\item{times}{Optional numeric vector of time points. Overrides \code{n_days}.
Defaults to \code{0:(n_days - 1)} when \code{NULL}.}

\item{method}{Integration method passed to \code{deSolve::ode()}.}

\item{...}{Currently unused.}
}
\value{
An object of class \code{"epi_model_predict"}.
}
\description{
Computes forward predictions from a fitted epidemic model by solving the
underlying ODE system using the estimated parameters.
}
\examples{
\dontrun{
sim <- simulate_epi(
  model = SIR_MODEL,
  n_days = 200,
  parms = c(beta = 0.30, gamma = 0.10),
  init = list(S = 1e6, I = 20, R = 0),
  obs = "poisson"
)
plot(sim)
inc_obs <- sim$incidence$inc
plot(seq_along(inc_obs) - 1, inc_obs,
     type = "l", xlab = "Day", ylab = "Incidence")
fit <- fit_epi_model(inc_obs,
                     loss = "logrmse",
                     model = SIRS_MODEL,
                     init = list(I0 = 6, N = 1e6))
init <- tail(sim$states, n = 1)[, -1]
pred <- predict(
  object = fit,
  n_days = 1000,
  init = init
)
plot(pred)
plot(sim)
plot(pred$states$time, pred$states$I)
plot(pred$states$I, col = "red", lty = 2)
lines(pred$states$S, col = "blue", lty = 2)
}

}
