% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_Z_list}
\alias{prepare_Z_list}
\title{Helper to prepare Z (4D array) for C++ consumption}
\usage{
prepare_Z_list(Z)
}
\arguments{
\item{Z}{The 4D array of exogenous covariates.}
}
\value{
A list of 3D arrays.
}
\description{
C++ (RcppArmadillo) handles 4D arrays best when passed as a list of 3D cubes.
We convert Z (m x m x q x T) into a list of q (m x m x T) cubes.
}
