% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium_demography.R
\name{prob_death_who_to_annual}
\alias{prob_death_who_to_annual}
\title{Smooth WHO life table death probabilites}
\usage{
prob_death_who_to_annual(probability_of_death, spar = 0.2)
}
\arguments{
\item{probability_of_death}{Column: "nqx - probability of dying between ages x and x+n"}

\item{spar}{Smoothing parameter. See \link[stats]{smooth.spline} for more details}
}
\value{
Smoothed probability of death
}
\description{
Life table raw data can be obtained from http://apps.who.int/gho/data/node.main.687?lang=en.
}
