% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_infection.R
\name{random_infection}
\alias{random_infection}
\title{Random infection of a matrix}
\usage{
random_infection(nrow = 20, ncol = 20, prob = 0.15)
}
\arguments{
\item{nrow}{An integer value of number of rows.}

\item{ncol}{An integer value of number of columns.}

\item{prob}{A float value of the probability of infection.}
}
\value{
A matrix of special class: SIRmatrix.
}
\description{
Random infection of a matrix
}
\examples{
random_infection()

random_infection(50, 50)

random_infection(25, 25, 0.15)
}
