% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM_closed_SIR.R
\name{rate_to_prob}
\alias{rate_to_prob}
\title{Convert a rate to a probability}
\usage{
rate_to_prob(rate, dt)
}
\arguments{
\item{rate}{Rate}

\item{dt}{timestep}
}
\value{
A probability
}
\description{
Convert a rate to a probability
}

