% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadAlevin.R
\name{ReadAlevinCsv}
\alias{ReadAlevinCsv}
\title{https://combine-lab.github.io/alevin-tutorial/2018/alevin-seurat/}
\usage{
ReadAlevinCsv(base.path = NULL)
}
\description{
https://combine-lab.github.io/alevin-tutorial/2018/alevin-seurat/
}
