% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{reinfection_model}
\alias{reinfection_model}
\title{Generate Reinfection Model}
\usage{
reinfection_model(time, state, parameters)
}
\arguments{
\item{time}{unused parameter, needed by deSolve}

\item{state}{state variable values, in model order (see return)}

\item{parameters}{model parameters}
}
\value{
list of values of all derivatives of model
}
\description{
This function generates the reinfection model for use with deSolve package
}
