% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{run_model}
\alias{run_model}
\title{run stochastic SIR model}
\usage{
run_model(tt, beta = 0.2, sigma = 0.1, S_ini = 1000, I_ini = 1, plot = FALSE)
}
\arguments{
\item{tt}{vector of times}

\item{beta}{rate of infection}

\item{sigma}{rate of recovery}

\item{S_ini}{initial number susceptible}

\item{I_ini}{initial number infected}
}
\description{
runs the stochastic SIR model for given times and parameters
}
