% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{run_model_deterministic}
\alias{run_model_deterministic}
\title{run deterministic SIR model}
\usage{
run_model_deterministic(
  tt,
  beta = 0.2,
  sigma = 0.1,
  S_ini = 1000,
  I_ini = 1,
  plot = FALSE
)
}
\arguments{
\item{tt}{vector of times}

\item{beta}{rate of infection}

\item{sigma}{rate of recovery}

\item{S_ini}{initial number susceptible}

\item{I_ini}{initial number infected}
}
\description{
runs the deterministic SIR model for given times and parameters
}
