% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_double}
\alias{sample_double}
\title{Sample single double from range.}
\usage{
sample_double(from, to)
}
\arguments{
\item{from}{Starting double (inclusive)}

\item{to}{End double (inclusive)}
}
\description{
Sample single double from range.
}
