% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_int}
\alias{sample_int}
\title{Sample single integer from range.}
\usage{
sample_int(from, to)
}
\arguments{
\item{from}{Starting integer (inclusive)}

\item{to}{End integer (inclusive)}
}
\description{
Sample single integer from range.
}
