% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{sir_c}
\alias{sir_c}
\title{SIR model with cumulative infections and incidence output}
\usage{
sir_c(time, state, parms)
}
\arguments{
\item{time}{Numeric scalar. Current time \eqn{t} at which derivatives are
evaluated (provided by \code{deSolve::ode()}).}

\item{state}{Named numeric vector of current state values. Must include
\code{S}, \code{I}, \code{R}, and \code{C}.}

\item{parms}{Named numeric vector of model parameters. Must include
\code{beta} and \code{gamma}.}
}
\value{
A list with two elements:
\describe{
\item{1}{A numeric vector of derivatives \code{c(dS, dI, dR, dC)} (in this order).}
\item{incidence}{A numeric scalar with the incidence \eqn{\lambda(t)} at \code{time}.}
}
}
\description{
Defines the right-hand side (derivatives) of an SIR compartmental model
extended with an auxiliary cumulative-infections state \code{C}. This function
is intended to be passed as the \code{func} argument to \code{deSolve::ode()}.
}
\details{
\subsection{State variables}{

The state vector \code{state} must contain (at least) the following named
components:
\describe{
\item{S}{Number of susceptible individuals at time \eqn{t}.}
\item{I}{Number of infectious (active) individuals at time \eqn{t}.}
\item{R}{Number of removed/recovered individuals at time \eqn{t}.}
\item{C}{Cumulative number of infections up to time \eqn{t}.}
}
}

\subsection{Parameters}{

The parameter vector \code{parms} must contain:
\describe{
\item{beta}{Transmission rate (per day).}
\item{gamma}{Recovery/removal rate (per day).}
}
}

\subsection{Model equations}{

The total population is computed as \eqn{N = S + I + R}. The incidence
(new infections per day) is:
\deqn{\lambda(t) = \beta \frac{S(t) I(t)}{N}}

and the ODE system is:
\deqn{\frac{dS}{dt} = -\lambda(t)}
\deqn{\frac{dI}{dt} = \lambda(t) - \gamma I(t)}
\deqn{\frac{dR}{dt} = \gamma I(t)}
\deqn{\frac{dC}{dt} = \lambda(t)}

The function returns a list with the derivatives and an additional named output
\code{incidence} equal to \eqn{\lambda(t)}, which will appear as an extra column
in the output of \code{deSolve::ode()}.
}
}
\examples{
library(deSolve)

pars <- c(beta = 0.30, gamma = 0.10)
init <- c(S = 999, I = 1, R = 0, C = 1)
times <- 0:160

out <- ode(y = init, times = times, func = sir_c, parms = pars, method = "lsoda")
head(out)

# Plot infectious prevalence I(t)
plot(out[, "time"], out[, "I"], type = "l", xlab = "Days", ylab = "I(t)")

# Plot incidence lambda(t)
plot(out[, "time"], out[, "incidence"], type = "l", xlab = "Days", ylab = "Incidence")

}
