% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_lines}
\alias{sir_lines}
\title{add output to existing plot from an SIR model run}
\usage{
sir_lines(res, ...)
}
\arguments{
\item{res}{object generated by running sir::run_model or sir::run_model_deterministic}

\item{...}{optional arguments to pass to base plot() function}
}
\description{
add output to existing plot from an SIR model run
}
