% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{SIR_MODEL}
\alias{SIR_MODEL}
\title{SIR epidemic model with cumulative infections}
\format{
An object of class \code{"epi_model"}.
}
\usage{
SIR_MODEL
}
\description{
An \code{epi_model} object representing a deterministic \strong{SIR}
(Susceptible–Infectious–Recovered) compartmental epidemic model. The model is
extended with an auxiliary state variable \code{C(t)} that tracks the cumulative
number of infections over time, and it provides the instantaneous incidence as
an additional model output.
}
\details{
\subsection{State variables}{

The model is defined in terms of the following state variables:
\describe{
\item{S(t)}{Number of susceptible individuals at time \eqn{t}.}
\item{I(t)}{Number of infectious (actively infected) individuals at time \eqn{t}.}
\item{R(t)}{Number of removed/recovered individuals at time \eqn{t}.}
\item{C(t)}{Cumulative number of infections up to time \eqn{t}.}
}

The total population size is given by
\deqn{N = S(t) + I(t) + R(t),}
which is conserved by the model dynamics.
}

\subsection{Parameters}{

The SIR model depends on the following parameters:
\describe{
\item{beta}{Transmission rate (per day).}
\item{gamma}{Recovery/removal rate (per day).}
}
}

\subsection{Model equations}{

New infections occur at rate
\deqn{\lambda(t) = \beta \frac{S(t)\, I(t)}{N}.}

The system of ordinary differential equations is:
\deqn{
\begin{aligned}
\frac{dS}{dt} &= -\lambda(t), \\
\frac{dI}{dt} &= \lambda(t) - \gamma I(t), \\
\frac{dR}{dt} &= \gamma I(t), \\
\frac{dC}{dt} &= \lambda(t).
\end{aligned}
}

The auxiliary state \eqn{C(t)} provides a continuous-time analogue of cumulative
incidence and is useful for linking the latent epidemic dynamics to
discrete-time observation models.
}

\subsection{Usage}{

This predefined model object is intended to be used with generic utilities
such as \code{\link{simulate_epi}} and model-fitting functions that operate
on \code{epi_model} objects.
}
}
\examples{
## Simulate a SIR epidemic without an observation model
sim <- simulate_epi(
  model = SIR_MODEL,
  n_days = 200,
  parms = c(beta = 0.3, gamma = 0.1),
  init_args = list(N = 1e6, I0 = 20, R0 = 0),
  obs = "none"
)

plot(sim$states$time, sim$states$I, type = "l",
     xlab = "Days", ylab = "I(t)",
     main = "SIR model: infectious individuals")

}
\seealso{
\code{\link{simulate_epi}}, \code{\link{new_epi_model}}
}
\keyword{datasets}
