% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir_plot}
\alias{sir_plot}
\title{plot output from an SIR model run}
\usage{
sir_plot(res)
}
\arguments{
\item{res}{object generated by running sir::run_model or sir::run_model_deterministic}
}
\description{
plot output from an SIR model run
}
