% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sir_test}
\alias{sir_test}
\title{Test demography}
\usage{
sir_test(N, days, substep, age_of_death, equilibrium_age, beta, sigma, i0)
}
\arguments{
\item{N}{Population size}

\item{days}{Simulation time (days)}

\item{substep}{Substeps per day}

\item{age_of_death}{Probability of age of death}

\item{equilibrium_age}{Equilibrium age distribution}

\item{beta}{Beta parameter}

\item{sigma}{Recovery rate}

\item{i0}{Initial number of infecteds}
}
\description{
Test demography
}
