% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{solve_system}
\alias{solve_system}
\title{Solve Disease Model}
\usage{
solve_system(
  type = "reinfection",
  parameters = NULL,
  initial_conditions = NULL,
  timesteps = NULL,
  path_num = NULL
)
}
\arguments{
\item{type}{model to solve; default reinfection, supported values
reinfection, complex, or basic}

\item{parameters}{system parameters}

\item{initial_conditions}{initial condition of the system}

\item{timesteps}{sequence of timesteps to sove the system over}

\item{path_num}{optional identification number to identify this path}
}
\value{
data.frame of time solutions of the system
}
\description{
This function utilizes the deSolve package to solve the systems of
differential equations specified in reiteration_model, complex_model, and
basic_model
}
