% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{step_infection}
\alias{step_infection}
\title{Create a randomly infected matrix.}
\usage{
step_infection(x = 50, z = 50, w = 0.1)
}
\arguments{
\item{x}{number of rows in grid}

\item{z}{number of columns in grid}

\item{w}{probability of initial infection}
}
\description{
Create a randomly infected matrix.
}
