% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{step_removed}
\alias{step_removed}
\title{Return a matrix with infected converted to removed/recovered cells.}
\usage{
step_removed(x)
}
\arguments{
\item{x}{a matrix with susceptible, infected, and removed values}
}
\description{
Return a matrix with infected converted to removed/recovered cells.
}
