% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{step_susceptible}
\alias{step_susceptible}
\title{Create a matrix with previous infections removed/recovered and calculate new infection spread.}
\usage{
step_susceptible(x, y = 0.2)
}
\arguments{
\item{x}{a matrix with susceptible, infected, and removed values}

\item{y}{probability of infecting a neighbor cell}
}
\description{
Create a matrix with previous infections removed/recovered and calculate new infection spread.
}
