% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.SIRmatrix.R
\name{summary.SIRmatrix}
\alias{summary.SIRmatrix}
\title{Summary statistics for a given SIRmatrix}
\usage{
\method{summary}{SIRmatrix}(object, ...)
}
\arguments{
\item{object}{A matrix of class SIRmatrix.}

\item{...}{Any other summary statistic parameters you'd want to include.}
}
\value{
A list continained all summary statistics.
}
\description{
Summary statistics for a given SIRmatrix
}
\examples{
summary(infect_corners())

summary(random_infection())
}
