% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_sample_int}
\alias{weighted_sample_int}
\title{Sample single integer from vector with probability weights.}
\usage{
weighted_sample_int(x, weights)
}
\arguments{
\item{x}{Vector of integers to sample from}

\item{weights}{Vector of weights}
}
\value{
Probability weighted sample
}
\description{
Sample single integer from vector with probability weights.
}
