% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_sample_min_int}
\alias{weighted_sample_min_int}
\title{Sample single integer from vector with probability weights conditional on
minimum lower bound.}
\usage{
weighted_sample_min_int(x, weights, lower)
}
\arguments{
\item{x}{Vector of integers to sample from (must be ordered)}

\item{weights}{Vector of weights}

\item{lower}{Lower value of x that must be exceeded}
}
\value{
Probability weighted sample
}
\description{
Sample single integer from vector with probability weights conditional on
minimum lower bound.
}
