% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{years_to_steps}
\alias{years_to_steps}
\title{Convert years to steps}
\usage{
years_to_steps(years, substep)
}
\arguments{
\item{years}{Number of years}

\item{substep}{Substeps per day}
}
\description{
Convert years to steps
}
