% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{AIC.sir}
\alias{AIC.sir}
\title{Calculate AIC}
\usage{
\method{AIC}{sir}(object, ..., k = 2)
}
\arguments{
\item{object}{A sir object}

\item{...}{Additional arguments for comparison with other models}

\item{k}{Penalty parameter (default 2 for AIC)}
}
\description{
Calculate AIC
}
