% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{attack_rate}
\alias{attack_rate}
\title{Attack rate}
\usage{
attack_rate(incidence)
}
\arguments{
\item{incidence}{Numeric vector giving the incidence curve.}
}
\value{
A numeric scalar giving the attack rate.
}
\description{
Computes the cumulative number of events as the sum of an incidence curve.
}
\details{
Let \eqn{\lambda(t)} denote the incidence function.
The attack rate is defined as
\deqn{
AR = \int_0^T \lambda(t)\, dt.
}

In discrete time, this is approximated by:
\deqn{
AR \approx \sum_t \lambda(t).
}
}
\examples{
inc <- c(1, 2, 3, 4)
attack_rate(inc)

}
