% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{BIC.sir}
\alias{BIC.sir}
\title{Calculate BIC}
\usage{
\method{BIC}{sir}(object, ...)
}
\arguments{
\item{object}{A sir object}

\item{...}{Additional arguments for comparison with other models}
}
\description{
Calculate BIC
}
