% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{coef.sir}
\alias{coef.sir}
\title{S3 Methods for SIR Objects}
\usage{
\method{coef}{sir}(object, ...)
}
\arguments{
\item{object}{A sir object}

\item{...}{Additional arguments (unused)}
}
\description{
Collection of S3 methods for objects of class "sir", providing standard model
interface functions including print, summary, plot, coef, fitted, residuals,
and model comparison tools.
}
