% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_construct_Walpha_design}
\alias{cpp_construct_Walpha_design}
\title{Construct Walpha design matrix for ALS update (theta, beta)}
\usage{
cpp_construct_Walpha_design(W, X, alpha)
}
\arguments{
\item{W}{(m x m x p) array of influence covariates.}

\item{X}{(m x m x T) array, typically lagged outcomes.}

\item{alpha}{(p x 1) vector of current alpha estimates.}
}
\value{
(m*m*T) x p matrix for the GLM design matrix.
}
\description{
Efficiently calculates and flattens the Walpha component of the design matrix.
Walpha[,,k,t] corresponds to (W*alpha) %*% X[,,t] %*% W[,,k]^T.
}
