% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_construct_Wbeta_design}
\alias{cpp_construct_Wbeta_design}
\title{Construct Wbeta design matrix for ALS update (theta, alpha)}
\usage{
cpp_construct_Wbeta_design(W, X, beta)
}
\arguments{
\item{W}{(m x m x p) array of influence covariates.}

\item{X}{(m x m x T) array, typically lagged outcomes.}

\item{beta}{(p x 1) vector of current beta estimates.}
}
\value{
(m*m*T) x p matrix for the GLM design matrix.
}
\description{
Efficiently calculates and flattens the Wbeta component of the design matrix.
Wbeta[,,k,t] corresponds to W[,,k] %*% X[,,t] %*% (W*beta)^T.
}
