% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{doubling_time_ts}
\alias{doubling_time_ts}
\title{Time-varying doubling time}
\usage{
doubling_time_ts(sim)
}
\arguments{
\item{sim}{An object of class \code{"sim_epi"}.}
}
\value{
A data.frame with columns \code{time} and \code{doubling_time}.
}
\description{
Computes the time-varying epidemic doubling time based on the instantaneous
growth rate.
}
\details{
Given the instantaneous growth rate \eqn{r(t)}, the doubling time is
\deqn{
T_d(t) = \frac{\log 2}{r(t)}.
}

When \eqn{r(t) \le 0}, the doubling time is set to infinity.

Requires a model flow named \code{"incidence"}.
}
\examples{
sim <- simulate_epi(
  model = SI_MODEL,
  times = 0:100,
  parms = c(beta = 0.3),
  init  = c(S = 999, I = 1)
)

head(doubling_time_ts(sim))

}
