% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eta_tab}
\alias{eta_tab}
\title{Calculate the bilinear predictor (eta)}
\usage{
eta_tab(tab, W, X, Z)
}
\arguments{
\item{tab}{Parameter vector (theta, alpha[-1], beta).}

\item{W}{(m x m x p) array of influence covariates.}

\item{X}{(m x m x T) array for the bilinear part.}

\item{Z}{(m x m x q x T) array of exogenous covariates.}
}
\value{
An (m x m x T) array of linear predictors.
}
\description{
Computes eta_{i,j,t} = Z_{i,j,t}^T theta + alpha^T X_{i,j,t} beta.
Uses optimized C++ functions for tensor products.
}
