% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{fitted.sir}
\alias{fitted.sir}
\title{Extract Fitted Values}
\usage{
\method{fitted}{sir}(object, ...)
}
\arguments{
\item{object}{A sir object}

\item{...}{Additional arguments (unused)}
}
\description{
Extract Fitted Values
}
