% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_api.R
\name{get_model}
\alias{get_model}
\title{Get registered model}
\usage{
get_model(name)
}
\arguments{
\item{name}{Character scalar with the model name to retrieve.}
}
\value{
An object of class \code{"epi_model"}.
}
\description{
Retrieves a model from the registry by name.
}
