% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_api.R
\name{get_model}
\alias{get_model}
\title{Get a registered model}
\usage{
get_model(name)
}
\arguments{
\item{name}{Character scalar with model name.}
}
\value{
An object of class "epi_model".
}
\description{
Get a registered model
}
