% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBM_closed_SIR.R
\name{ibm_closed_SIR}
\alias{ibm_closed_SIR}
\title{Closed SIR (IBM)}
\usage{
ibm_closed_SIR(sigma = 0.4, beta = 0.5, N = 1000, i0 = 1,
  t_final = 100, dt = 0.01)
}
\arguments{
\item{sigma}{Recovery rate (1 / mean duration of infection)}

\item{beta}{Transmission parameter}

\item{N}{Population size}

\item{i0}{Number of infected people at time = 0}

\item{t_final}{End time}

\item{dt}{time step}
}
\value{
IBM closed SIR model output
}
\description{
Closed SIR (IBM)
}

