% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{initial_growth_rate}
\alias{initial_growth_rate}
\title{Initial exponential growth rate}
\usage{
initial_growth_rate(sim, n = 7)
}
\arguments{
\item{sim}{An object of class \code{"sim_epi"}.}

\item{n}{Integer. Number of initial time points.}
}
\value{
A numeric scalar giving the initial exponential growth rate.
}
\description{
Estimates the early exponential growth rate of the epidemic by fitting a
log-linear model to the initial segment of the incidence curve.
}
\details{
During the early epidemic phase, incidence is assumed to grow exponentially:
\deqn{
\lambda(t) \approx C e^{r t},
}
where \eqn{r} is the exponential growth rate.

Taking logarithms yields
\deqn{
\log \lambda(t) = \log C + r t.
}

Requires the epidemiological role \code{"incidence"}.
}
\examples{
sim <- simulate_epi(
  model = SI_MODEL,
  times = 0:50,
  parms = c(beta = 0.4),
  init  = c(S = 999, I = 1)
)

initial_growth_rate(sim, n = 10)

}
