% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{logLik.sir}
\alias{logLik.sir}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{sir}(object, ...)
}
\arguments{
\item{object}{A sir object}

\item{...}{Additional arguments (unused)}
}
\description{
Extract Log-Likelihood
}
