% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_closed_SIR.R
\name{odin_closed_sir}
\alias{odin_closed_sir}
\title{Closed SIR (ODIN)}
\usage{
odin_closed_sir(N = 1000, days = 100, substep = 1, sigma = 2,
  beta = 6, i0 = 1)
}
\arguments{
\item{N}{Population size}

\item{substep}{Number of substeps per day}

\item{sigma}{Recovery rate (1 / mean duration of infection)}

\item{beta}{Transmission parameter}

\item{i0}{Number of infected people at time = 0}

\item{Days}{Number of days}
}
\value{
Odin closed SIR model output
}
\description{
Closed SIR (ODIN)
}
