% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_closed_SIR.R
\name{odin_closed_sir}
\alias{odin_closed_sir}
\title{Closed SIR (ODIN)}
\usage{
odin_closed_sir(sigma = 2, beta = 6, N = 1000, i0 = 1,
  t_final = 100, dt = 0.01)
}
\arguments{
\item{sigma}{Recovery rate (1 / mean duration of infection)}

\item{beta}{Transmission parameter}

\item{N}{Population size}

\item{i0}{Number of infected people at time = 0}

\item{t_final}{End time}

\item{dt}{time step}
}
\value{
Odin closed SIR model output
}
\description{
Closed SIR (ODIN)
}
