% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_closed_SIR.R
\name{odin_closed_sir}
\alias{odin_closed_sir}
\title{Closed SIR (ODIN)}
\usage{
odin_closed_sir(sigma = 2, beta = 4, init = c(1000, 1, 0), t = seq(0,
  10, length.out = 2000))
}
\arguments{
\item{sigma}{Recovery rate (1 / mean duration of infection)}

\item{beta}{Transmission parameter (contact rate * probability of infection upon contact)}

\item{init}{Vector of initial values (S, I, R)}
}
\value{
An odin model
}
\description{
Closed SIR (ODIN)
}
\examples{
m1 <- odin_closed_sir()
plot(m1$S ~ m1$t,  t = "l", ylim = c(0, 1001))
lines(m1$I ~ m1$t, col = "red")
lines(m1$R ~ m1$t,  col = "blue")
}

