% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_sir_wrapper.R
\name{open_sir}
\alias{open_sir}
\title{Open sIR}
\usage{
open_sir(N = 1000, days = 100, substep = 1, sigma = 0.003,
  beta = 0.05, i0 = 10, aod = age_of_death, eq = equilibrium_age)
}
\arguments{
\item{N}{Number of individuals}

\item{days}{Number of days}

\item{substep}{Steps per day}

\item{sigma}{Recovery rate}

\item{beta}{Infection rate}

\item{i0}{Initial number of infected}

\item{aod}{Distribution of age of death}

\item{eq}{Equilibrium age distribution}
}
\value{
A processed data.frame of model output
}
\description{
Open SIR IBM, implemented internally in c++
}
