% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{peak_incidence}
\alias{peak_incidence}
\title{Peak incidence}
\usage{
peak_incidence(sim)
}
\arguments{
\item{sim}{An object of class \code{"sim_epi"}.}
}
\value{
A named list with elements:
\itemize{
\item \code{time}: Time at which peak incidence occurs.
\item \code{value}: Maximum incidence value.
}
}
\description{
Computes the maximum value of the incidence curve and the time at which
it occurs.

This metric requires the epidemiological role \code{"incidence"}.
}
\examples{
sim <- simulate_epi(
  model = SI_MODEL,
  times = 0:100,
  parms = c(beta = 0.3),
  init  = c(S = 999, I = 1)
)

peak_incidence(sim)

}
