% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{peak_prevalence}
\alias{peak_prevalence}
\title{Peak prevalence}
\usage{
peak_prevalence(sim)
}
\arguments{
\item{sim}{An object of class \code{"sim_epi"}.}
}
\value{
A numeric scalar giving the peak prevalence.
}
\description{
Computes the maximum number of infectious individuals during the epidemic.
}
\details{
Let \eqn{I(t)} denote the number of infectious individuals over time.
The peak prevalence is defined as
\deqn{
I_{\max} = \max_t I(t).
}

This metric requires the epidemiological role \code{"infectious"}.
}
\examples{
sim <- simulate_epi(
  model = SIR_MODEL,
  times = 0:160,
  parms = c(beta = 0.3, gamma = 0.1),
  init  = c(S = 999, I = 1, R = 0)
)

peak_prevalence(sim)

}
