% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos.R
\name{plot.sim_epi}
\alias{plot.sim_epi}
\title{Plot a simulated epidemic}
\usage{
\method{plot}{sim_epi}(x, what = c("states", "incidence"), ...)
}
\arguments{
\item{x}{Object of class \code{"sim_epi"} as returned by
\code{\link{simulate_epi}}.}

\item{what}{Character. Type of plot to produce. Either \code{"states"} for the
compartmental trajectories (S, I, R) or \code{"incidence"} for the observed
incidence time series.}

\item{...}{Additional graphical parameters passed to base plotting functions
such as \code{\link{plot}} and \code{\link{lines}}.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Plot method for objects of class \code{"sim_epi"} as returned by
\code{\link{simulate_epi}}. The function can display either the compartmental
state trajectories (S, I, R) or the observed incidence time series, depending
on the value of the \code{what} argument.
}
\details{
When \code{what = "states"}, the function plots the time evolution of the
susceptible (S), infectious (I), and recovered (R) compartments on the same
figure. When \code{what = "incidence"}, the function plots the observed
incidence counts generated by the observation model.

This method is automatically dispatched when calling \code{plot()} on an
object of class \code{"sim_epi"}.
}
\examples{
sim <- simulate_epi(n_days = 200, model = "sir", seed = 1)

# Plot S, I, R trajectories
plot(sim)

# Plot observed incidence
plot(sim, what = "incidence")

}
\seealso{
\code{\link{simulate_epi}}, \code{\link{summary.epi_sim}}
}
