% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_sir_wrapper.R
\name{post}
\alias{post}
\title{Post processing wrapper for c++ output}
\usage{
post(out, days, substep)
}
\arguments{
\item{out}{Raw model output}

\item{days}{Number of days}

\item{substep}{Steps per day}
}
\value{
Processed c++ output
}
\description{
Post processing wrapper for c++ output
}
