% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{predict.sir}
\alias{predict.sir}
\title{Predict Method for SIR Objects}
\usage{
\method{predict}{sir}(object, newdata = NULL, type = c("link", "response"), ...)
}
\arguments{
\item{object}{A sir object}

\item{newdata}{New data for prediction}

\item{type}{Type of prediction}

\item{...}{Additional arguments}
}
\description{
Predict Method for SIR Objects
}
