% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{print.sir}
\alias{print.sir}
\title{Simplified Print Method for SIR Objects}
\usage{
\method{print}{sir}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A sir object}

\item{digits}{Number of digits to print}

\item{...}{Additional arguments}
}
\description{
Simplified Print Method for SIR Objects
}
