% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{print.summary.sir}
\alias{print.summary.sir}
\title{Print Summary of SIR Model}
\usage{
\method{print}{summary.sir}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{A summary.sir object}

\item{digits}{Number of significant digits to print}

\item{signif.stars}{Logical, whether to show significance stars}

\item{...}{Additional arguments (unused)}
}
\description{
Print Summary of SIR Model
}
