% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{register_epi_model}
\alias{register_epi_model}
\title{Register a new epidemiological model}
\usage{
register_epi_model(name, model, overwrite = FALSE)
}
\arguments{
\item{name}{Model name}

\item{model}{An epi_model object}

\item{overwrite}{Overwrite existing model?}
}
\description{
Register a new epidemiological model
}
\examples{
## ---------------------------------------------------------------
## Define and register a new epidemiological model (SIRD)
## ---------------------------------------------------------------

sird_rhs <- function(time, state, parms) {
  with(as.list(c(state, parms)), {
    N <- S + I + R
    lambda <- beta * S * I / N

    dS <- -lambda
    dI <-  lambda - gamma * I - mu * I
    dR <-  gamma * I
    dD <-  mu * I

    list(
      c(dS, dI, dR, dD),
      incidence = lambda
    )
  })
}

SIRD_MODEL <- new_epi_model(
  name        = "SIRD",
  rhs         = sird_rhs,
  state_names = c("S", "I", "R", "D"),
  par_names   = c("beta", "gamma", "mu"),
  defaults    = c(beta = 0.3, gamma = 0.1, mu = 0.01),
  init        = c(S = 1e6, I = 10, R = 0, D = 0),
  outputs     = c("S", "I", "R", "D", "incidence")
)

register_epi_model("SIRD", SIRD_MODEL)

## Simulate using explicit initial conditions
sim <- simulate_epi(
  model = get_epi_model("SIRD"),
  times = 0:200,
  parms = c(beta = 0.25, gamma = 0.1, mu = 0.02),
  init  = get_epi_model("SIRD")$init
)

plot(sim)
plot(sim, what = "incidence")
}
