% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry_api.R
\name{register_epi_model}
\alias{register_epi_model}
\title{Register an epidemiological model}
\usage{
register_epi_model(model)
}
\arguments{
\item{model}{An object of class "epi_model".}
}
\description{
Register an epidemiological model
}
