% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{residuals.sir}
\alias{residuals.sir}
\title{Extract Model Residuals}
\usage{
\method{residuals}{sir}(object, type = c("deviance", "pearson", "response"), ...)
}
\arguments{
\item{object}{A sir object}

\item{type}{Type of residuals: "deviance", "pearson", or "response"}

\item{...}{Additional arguments (unused)}
}
\description{
Extract Model Residuals
}
