% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_open_sir_wrapper.R
\name{set_time}
\alias{set_time}
\title{Set time}
\usage{
set_time(out, step)
}
\arguments{
\item{out}{Raw model output}
}
\value{
Model output with time columns added
}
\description{
Adds model time and calendar (day, month, year) time to c++ model output
}
