% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir}
\alias{sir}
\title{Create an SIR simulation with given arguments.}
\usage{
sir(
  x = step_infection(nrow, ncol, p0),
  nrow = 50,
  ncol = 50,
  p = 0.2,
  p0 = 0.1,
  plot = FALSE,
  time_per_frame = 0.2,
  ...
)
}
\arguments{
\item{nrow}{number of rows in grid}

\item{ncol}{number of columns in grid}

\item{p}{probability of infecting a neighbor cell}

\item{p0}{initial probability of population to be infected}

\item{plot}{set to TRUE to display plot of matrix}

\item{time_per_frame}{time interval between each infection step}
}
\description{
Create an SIR simulation with given arguments.
}
