% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{SIRS_MODEL}
\alias{SIRS_MODEL}
\title{SIRS epidemic model with waning immunity}
\format{
An object of class \code{"epi_model"}.
}
\usage{
SIRS_MODEL
}
\description{
An \code{epi_model} object representing a deterministic \strong{SIRS}
(Susceptible–Infectious–Recovered–Susceptible) compartmental epidemic model
with waning immunity. Individuals who recover from infection lose immunity
at rate \code{omega} and return to the susceptible compartment.
}
\details{
\subsection{State variables}{

The model is defined in terms of the following state variables:
\describe{
\item{S(t)}{Number of susceptible individuals at time \eqn{t}.}
\item{I(t)}{Number of infectious (actively infected) individuals at time \eqn{t}.}
\item{R(t)}{Number of recovered (temporarily immune) individuals at time \eqn{t}.}
}

The total population size is given by
\deqn{N = S(t) + I(t) + R(t),}
which is conserved by the model dynamics.
}

\subsection{Parameters}{

The SIRS model depends on the following parameters:
\describe{
\item{beta}{Transmission rate (per day).}
\item{gamma}{Recovery/removal rate (per day).}
\item{omega}{Rate of waning immunity from \code{R} back to \code{S} (per day).}
}
}

\subsection{Model equations}{

New infections occur at rate
\deqn{\lambda(t) = \beta \frac{S(t)\, I(t)}{N}.}

The system of ordinary differential equations is:
\deqn{
\begin{aligned}
\frac{dS}{dt} &= -\lambda(t) + \omega R(t), \\
\frac{dI}{dt} &= \lambda(t) - \gamma I(t), \\
\frac{dR}{dt} &= \gamma I(t) - \omega R(t). \\
\end{aligned}
}
}

\subsection{Usage}{

This predefined model object is intended to be used with generic utilities
such as \code{\link{simulate_epi}} and model-fitting functions that operate
on \code{epi_model} objects.
}
}
\examples{
## Simulate a SIRS epidemic without an observation model
sim <- simulate_epi(
  model = SIRS_MODEL,
  n_days = 200,
  parms = c(beta = 0.3, gamma = 0.1, omega = 0.02),
  init_args = list(N = 1e6, I0 = 20, R0 = 0),
  obs = "none"
)

plot(sim$states$time, sim$states$I, type = "l",
     xlab = "Days", ylab = "I(t)",
     main = "SIRS model: infectious individuals")

}
\seealso{
\code{\link{simulate_epi}}, \code{\link{new_epi_model}}
}
\keyword{datasets}
