% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{summary.sir}
\alias{summary.sir}
\title{Summary Method for SIR Objects}
\usage{
\method{summary}{sir}(object, ...)
}
\arguments{
\item{object}{A sir object}

\item{...}{Additional arguments}
}
\value{
An object of class "summary.sir"
}
\description{
Summary Method for SIR Objects
}
