% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{time_to_peak}
\alias{time_to_peak}
\title{Time to peak incidence}
\usage{
time_to_peak(sim)
}
\arguments{
\item{sim}{An object of class \code{"sim_epi"}.}
}
\value{
A numeric scalar giving the time to peak incidence.
}
\description{
Returns the time at which incidence reaches its maximum.

This is a convenience wrapper around \code{peak_incidence()}.
}
\examples{
sim <- simulate_epi(
  model = SI_MODEL,
  times = 0:100,
  parms = c(beta = 0.3),
  init  = c(S = 999, I = 1)
)

time_to_peak(sim)

}
