% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{time_to_peak}
\alias{time_to_peak}
\title{Time to peak incidence}
\usage{
time_to_peak(incidence, time)
}
\arguments{
\item{incidence}{Numeric vector giving the incidence curve.}

\item{time}{Numeric vector of time points.}
}
\value{
A numeric scalar giving the time to peak incidence.
}
\description{
Returns the time at which incidence reaches its maximum.
}
\examples{
inc  <- c(1, 4, 6, 3)
time <- 0:3
time_to_peak(inc, time)

}
