% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir-methods.R
\name{vcov.sir}
\alias{vcov.sir}
\title{Extract Variance-Covariance Matrix}
\usage{
\method{vcov}{sir}(object, ...)
}
\arguments{
\item{object}{A sir object}

\item{...}{Additional arguments (unused)}
}
\description{
Extract Variance-Covariance Matrix
}
